MVP Card commit
This is the initial commit for Card + Player as we approach MVP. After lots of testing and bug fixing I have found a way to implement Cards by communicating between C++ and Blueprint classes. The next step is to begin implementing communication with other modules and to setup MVP ready functionality.